requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
torch==2.4.0
torchvision==0.17.0
Pillow==10.0.1
Werkzeug==2.2.3
